Componentize listings/show posted popup and head
